// src/app/models/utility.model.ts
export interface EventUtilitySuccess {
  id: number;
  name: string;
}

export interface CategoryUtilitySuccess {
  id: number;
  category_name: string;
}

export interface ProvinceUtilitySuccess {
  id: number;
  province_name: string;
}

export interface RoleUtilitySuccess {
  id: number;
  role_name: string;
}

export interface TicketUtilitySuccess {
  id: number;
  name: string;
}